+++ projects/packages/videopress/declarations.d.ts
@@ -2,10 +2,5 @@ declare module '*.png';
 declare module '*.svg';
 declare module '*.jpeg';
 declare module '*.jpg';
-
-export declare global {
-	interface Window {
-		// eslint-disable-next-line @typescript-eslint/no-explicit-any
-		wp: { media: any };
-	}
-}
+declare module '*.scss';
+declare module '*.mdx';