+++ projects/packages/videopress/src/class-data.php
@@ -196,6 +196,10 @@ class Data {
 				$media_details      = $video['media_details'];
 				$jetpack_videopress = $video['jetpack_videopress'];
 
+				// Check if video is already uploaded to VideoPress.
+				$uploader                  = new Uploader( $id );
+				$is_uploaded_to_videopress = $uploader->is_uploaded();
+
 				$upload_date = $video['date'];
 				$url         = $video['source_url'];
 
@@ -203,20 +207,21 @@ class Data {
 				$description = $jetpack_videopress['description'];
 				$caption     = $jetpack_videopress['caption'];
 
-				$width    = $media_details['width'];
-				$height   = $media_details['height'];
-				$duration = $media_details['length'];
+				$width    = isset( $media_details['width'] ) ? $media_details['width'] : null;
+				$height   = isset( $media_details['height'] ) ? $media_details['height'] : null;
+				$duration = isset( $media_details['length'] ) ? $media_details['length'] : null;
 
 				return array(
-					'id'          => $id,
-					'title'       => $title,
-					'description' => $description,
-					'caption'     => $caption,
-					'width'       => $width,
-					'height'      => $height,
-					'url'         => $url,
-					'uploadDate'  => $upload_date,
-					'duration'    => $duration,
+					'id'                     => $id,
+					'title'                  => $title,
+					'description'            => $description,
+					'caption'                => $caption,
+					'width'                  => $width,
+					'height'                 => $height,
+					'url'                    => $url,
+					'uploadDate'             => $upload_date,
+					'duration'               => $duration,
+					'isUploadedToVideoPress' => $is_uploaded_to_videopress,
 				);
 			},
 			$local_videos_data['videos']
@@ -242,7 +247,7 @@ class Data {
 				$privacy_setting = $jetpack_videopress['privacy_setting'];
 
 				$original      = $videopress_media_details['original'];
-				$poster        = $videopress_media_details['poster'];
+				$poster        = $privacy_setting !== 1 ? $videopress_media_details['poster'] : null;
 				$upload_date   = $videopress_media_details['upload_date'];
 				$duration      = $videopress_media_details['duration'];
 				$is_private    = $videopress_media_details['is_private'];
@@ -250,7 +255,7 @@ class Data {
 				$finished      = $videopress_media_details['finished'];
 				$files         = $videopress_media_details['files'];
 
-				if ( isset( $files['dvd']['original_img'] ) ) {
+				if ( isset( $files['dvd']['original_img'] ) && $privacy_setting !== 1 ) {
 					$thumbnail = $file_url_base['https'] . $files['dvd']['original_img'];
 				} else {
 					$thumbnail = null;