+++ projects/packages/videopress/src/class-wpcom-rest-api-v2-endpoint-videopress.php
@@ -159,6 +159,19 @@ class WPCOM_REST_API_V2_Endpoint_VideoPress extends WP_REST_Controller {
 				},
 			)
 		);
+
+		// Playback Token Route
+		register_rest_route(
+			$this->namespace,
+			$this->rest_base . '/playback-jwt/(?P<video_guid>\w+)',
+			array(
+				'methods'             => \WP_REST_Server::EDITABLE,
+				'callback'            => array( $this, 'videopress_playback_jwt' ),
+				'permission_callback' => function () {
+					return current_user_can( 'read' );
+				},
+			)
+		);
 	}
 
 	/**
@@ -306,6 +319,54 @@ class WPCOM_REST_API_V2_Endpoint_VideoPress extends WP_REST_Controller {
 	}
 
 	/**
+	 * Endpoint for generating a VideoPress Playback JWT
+	 *
+	 * @param WP_REST_Request $request the request object.
+	 * @return WP_Rest_Response - The response object.
+	 */
+	public static function videopress_playback_jwt( $request ) {
+		$has_connected_owner = Data::has_connected_owner();
+		if ( ! $has_connected_owner ) {
+			return rest_ensure_response(
+				new WP_Error(
+					'owner_not_connected',
+					'User not connected.',
+					array(
+						'code'        => 503,
+						'connect_url' => Admin_UI::get_admin_page_url(),
+					)
+				)
+			);
+		}
+
+		$blog_id = Data::get_blog_id();
+		if ( ! $blog_id ) {
+			return rest_ensure_response(
+				new WP_Error( 'site_not_registered', 'Site not registered.', 503 )
+			);
+		}
+
+		try {
+			$video_guid = $request->get_param( 'video_guid' );
+			$token      = VideoPressToken::videopress_playback_jwt( $video_guid );
+			$status     = 200;
+			$data       = array(
+				'playback_token' => $token,
+			);
+		} catch ( \Exception $e ) {
+			$status = 500;
+			$data   = array(
+				'error' => $e->getMessage(),
+			);
+
+		}
+
+		return rest_ensure_response(
+			new WP_REST_Response( $data, $status )
+		);
+	}
+
+	/**
 	 * Updates attachment meta and video metadata via the WPCOM REST API.
 	 *
 	 * @param WP_REST_Request $request the request object.
@@ -434,6 +495,11 @@ class WPCOM_REST_API_V2_Endpoint_VideoPress extends WP_REST_Controller {
 				if ( isset( $json_params['caption'] ) ) {
 					$meta['videopress']['caption'] = $post_excerpt;
 					$should_update_meta            = true;
+				}
+
+				if ( isset( $json_params['poster'] ) ) {
+					$meta['videopress']['poster'] = $json_params['poster'];
+					$should_update_meta           = true;
 				}
 
 				if ( isset( $json_params['allow_download'] ) ) {