+++ projects/packages/videopress/src/client/admin/components/admin-page/libraries.tsx
@@ -2,6 +2,7 @@
  * External dependencies
  */
 import { Button, Text, useBreakpointMatch } from '@automattic/jetpack-components';
+import { createInterpolateElement } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
 import { grid, formatListBullets } from '@wordpress/icons';
 import classnames from 'classnames';
@@ -10,7 +11,7 @@ import { useHistory } from 'react-router-dom';
 /**
  * Internal dependencies
  */
-import useVideos, { useLocalVideos } from '../../hooks/use-videos';
+import useVideos from '../../hooks/use-videos';
 import { SearchInput } from '../input';
 import { ConnectLocalPagination, ConnectPagination } from '../pagination';
 import { FilterButton, ConnectFilterSection } from '../video-filter';
@@ -20,7 +21,7 @@ import styles from './styles.module.scss';
 /**
  * Types
  */
-import { VideoLibraryProps } from './types';
+import { LocalLibraryProps, VideoLibraryProps } from './types';
 
 const LIBRARY_TYPE_LOCALSORAGE_KEY = 'videopress-library-type';
 
@@ -104,6 +105,7 @@ const VideoLibraryWrapper = ( {
 
 export const VideoPressLibrary = ( { videos, totalVideos, loading }: VideoLibraryProps ) => {
 	const history = useHistory();
+	const { search } = useVideos();
 
 	const libraryTypeFromLocalStorage = localStorage.getItem(
 		LIBRARY_TYPE_LOCALSORAGE_KEY
@@ -127,6 +129,23 @@ export const VideoPressLibrary = ( { videos, totalVideos, loading }: VideoLibrar
 		history.push( `/video/${ video?.id }/edit` );
 	};
 
+	const library =
+		libraryType === LibraryType.Grid ? (
+			<VideoGrid
+				videos={ videos }
+				onVideoDetailsClick={ handleClickEditDetails }
+				loading={ loading }
+				count={ uploading ? videos.length : 6 }
+			/>
+		) : (
+			<VideoList
+				videos={ videos }
+				onVideoDetailsClick={ handleClickEditDetails }
+				hidePlays
+				loading={ loading }
+			/>
+		);
+
 	return (
 		<VideoLibraryWrapper
 			totalVideos={ totalVideos }
@@ -134,47 +153,47 @@ export const VideoPressLibrary = ( { videos, totalVideos, loading }: VideoLibrar
 			libraryType={ libraryType }
 			title={ __( 'Your VideoPress library', 'jetpack-videopress-pkg' ) }
 		>
-			{ libraryType === LibraryType.Grid ? (
-				<VideoGrid
-					videos={ videos }
-					onVideoDetailsClick={ handleClickEditDetails }
-					loading={ loading }
-					count={ uploading ? videos.length : 6 }
-				/>
+			{ videos.length > 0 || loading ? (
+				library
 			) : (
-				<VideoList
-					videos={ videos }
-					onVideoDetailsClick={ handleClickEditDetails }
-					hidePlays
-					loading={ loading }
-				/>
+				<Text>
+					{ createInterpolateElement(
+						sprintf(
+							/* translators: placeholder is the search term */
+							__( 'No videos match your search for <em>%s</em>.', 'jetpack-videopress-pkg' ),
+							search
+						),
+						{
+							em: <em className={ styles[ 'query-no-results' ] } />,
+						}
+					) }
+				</Text>
 			) }
 			<ConnectPagination className={ styles.pagination } />
 		</VideoLibraryWrapper>
 	);
 };
 
-export const LocalLibrary = ( { videos, totalVideos, loading }: VideoLibraryProps ) => {
+export const LocalLibrary = ( {
+	videos,
+	totalVideos,
+	loading,
+	uploading,
+	onUploadClick,
+}: LocalLibraryProps ) => {
 	return (
 		<VideoLibraryWrapper
 			totalVideos={ totalVideos }
 			hideFilter
 			title={ __( 'Local videos', 'jetpack-videopress-pkg' ) }
 		>
-			<LocalVideoList videos={ videos } loading={ loading } />
+			<LocalVideoList
+				videos={ videos }
+				loading={ loading }
+				onActionClick={ onUploadClick }
+				uploading={ uploading }
+			/>
 			<ConnectLocalPagination className={ styles.pagination } />
 		</VideoLibraryWrapper>
-	);
-};
-
-export const ConnectLocalLibrary = () => {
-	const { items: videos, uploadedLocalVideoCount, isFetching } = useLocalVideos();
-
-	return (
-		<LocalLibrary
-			videos={ videos }
-			totalVideos={ uploadedLocalVideoCount }
-			loading={ isFetching }
-		/>
 	);
 };