+++ projects/packages/videopress/src/client/admin/components/edit-video-details/index.tsx
@@ -11,12 +11,14 @@ import {
 	useBreakpointMatch,
 } from '@automattic/jetpack-components';
 import { __ } from '@wordpress/i18n';
-import { Icon, chevronRightSmall } from '@wordpress/icons';
+import { Icon, chevronRightSmall, arrowLeft } from '@wordpress/icons';
 import classnames from 'classnames';
+import { useEffect } from 'react';
 import { useHistory, Prompt } from 'react-router-dom';
 /**
  * Internal dependencies
  */
+import { Link } from 'react-router-dom';
 import { VideoPlayer } from '../../../components/video-frame-selector';
 import useUnloadPrevent from '../../hooks/use-unload-prevent';
 import Input from '../input';
@@ -64,6 +66,19 @@ const Header = ( {
 					</Button>
 				</div>
 			</div>
+		</div>
+	);
+};
+
+const GoBackLink = () => {
+	const history = useHistory();
+
+	return (
+		<div className={ styles[ 'back-link' ] }>
+			<Link to="#" className={ styles.link } onClick={ () => history.push( '/' ) }>
+				<Icon icon={ arrowLeft } className={ styles.icon } />
+				{ __( 'Go back', 'jetpack-videopress-pkg' ) }
+			</Link>
 		</div>
 	);
 };
@@ -142,15 +157,20 @@ const EditVideoDetails = () => {
 		title,
 		description,
 		caption,
+		// Playback Token
+		playbackToken,
+		isFetchingPlaybackToken,
 		// Page State/Actions
 		hasChanges,
 		updating,
+		updated,
 		isFetching,
 		handleSaveChanges,
 		// Metadata
 		setTitle,
 		setDescription,
 		setCaption,
+		processing,
 		// Poster Image
 		useVideoAsThumbnail,
 		selectedTime,
@@ -170,25 +190,39 @@ const EditVideoDetails = () => {
 	);
 
 	useUnloadPrevent( {
-		shouldPrevent: hasChanges,
+		shouldPrevent: hasChanges && ! updated,
 		message: unsavedChangesMessage,
 	} );
 
+	const history = useHistory();
+
+	useEffect( () => {
+		if ( updated === true ) {
+			history.push( '/' );
+		}
+	}, [ updated ] );
+
+	// We may need the playback token on the video URL as well
+	const videoUrl = playbackToken ? `${ url }?metadata_token=${ playbackToken }` : url;
+
 	let thumbnail: string | JSX.Element = posterImage;
+
 	if ( posterImageSource === 'video' && useVideoAsThumbnail ) {
-		thumbnail = <VideoPlayer src={ url } currentTime={ selectedTime } />;
+		thumbnail = <VideoPlayer src={ videoUrl } currentTime={ selectedTime } />;
 	} else if ( posterImageSource === 'upload' ) {
 		thumbnail = libraryAttachment.url;
 	}
 
+	const isFetchingData = isFetching || isFetchingPlaybackToken;
+
 	return (
 		<>
-			<Prompt when={ hasChanges } message={ unsavedChangesMessage } />
+			<Prompt when={ hasChanges && ! updated } message={ unsavedChangesMessage } />
 
 			{ frameSelectorIsOpen && (
 				<VideoThumbnailSelectorModal
 					handleCloseSelectFrame={ handleCloseSelectFrame }
-					url={ url }
+					url={ videoUrl }
 					handleVideoFrameSelected={ handleVideoFrameSelected }
 					selectedTime={ selectedTime }
 					handleConfirmFrame={ handleConfirmFrame }
@@ -198,11 +232,14 @@ const EditVideoDetails = () => {
 			<AdminPage
 				moduleName={ __( 'Jetpack VideoPress', 'jetpack-videopress-pkg' ) }
 				header={
-					<Header
-						onSaveChanges={ handleSaveChanges }
-						saveDisabled={ ! hasChanges }
-						saveLoading={ updating }
-					/>
+					<>
+						<GoBackLink />
+						<Header
+							onSaveChanges={ handleSaveChanges }
+							saveDisabled={ ! hasChanges }
+							saveLoading={ updating }
+						/>
+					</>
 				}
 			>
 				<AdminSection>
@@ -215,14 +252,15 @@ const EditVideoDetails = () => {
 								onChangeDescription={ setDescription }
 								caption={ caption ?? '' }
 								onChangeCaption={ setCaption }
-								loading={ isFetching }
+								loading={ isFetchingData }
 							/>
 						</Col>
 						<Col sm={ 4 } md={ 8 } lg={ { start: 9, end: 12 } }>
 							<VideoThumbnail
-								thumbnail={ isFetching ? <Placeholder height={ 200 } /> : thumbnail }
+								thumbnail={ isFetchingData ? <Placeholder height={ 200 } /> : thumbnail }
 								duration={ duration }
 								editable
+								processing={ processing }
 								onSelectFromVideo={ handleOpenSelectFrame }
 								onUploadImage={ selectPosterImageFromLibrary }
 							/>
@@ -230,7 +268,7 @@ const EditVideoDetails = () => {
 								filename={ filename ?? '' }
 								uploadDate={ uploadDate ?? '' }
 								src={ url ?? '' }
-								loading={ isFetching }
+								loading={ isFetchingData }
 							/>
 						</Col>
 					</Container>