+++ projects/packages/videopress/src/client/admin/components/video-row/index.tsx
@@ -1,11 +1,16 @@
+/**
+ * External dependencies
+ */
 import { Text, Button, useBreakpointMatch } from '@automattic/jetpack-components';
 import { dateI18n } from '@wordpress/date';
 import { sprintf, __ } from '@wordpress/i18n';
 import { Icon, chevronDown, chevronUp } from '@wordpress/icons';
 import classNames from 'classnames';
 import { useState, useRef } from 'react';
-import privacy from '../../../components/icons/crossed-eye-icon';
-import { VIDEO_PRIVACY_LEVELS, VIDEO_PRIVACY_LEVEL_PRIVATE } from '../../../state/constants';
+/**
+ * Internal dependencies
+ */
+import privateIcon from '../../../components/icons/crossed-eye-icon';
 import useVideo from '../../hooks/use-video';
 import Checkbox from '../checkbox';
 import Placeholder from '../placeholder';
@@ -13,6 +18,9 @@ import { ConnectVideoQuickActions } from '../video-quick-actions';
 import VideoThumbnail from '../video-thumbnail';
 import StatsBase from './stats';
 import styles from './style.module.scss';
+/**
+ * Types
+ */
 import { VideoRowProps } from './types';
 
 const millisecondsToMinutesAndSeconds = ( milliseconds?: number ) => {
@@ -49,7 +57,7 @@ const Stats = ( {
 			<span>{ isPrivate ? privateLabel : publicLabel }</span>
 		</>
 	) : (
-		<>{ isPrivate && <Icon icon={ privacy } /> }</>
+		isPrivate && <Icon icon={ privateIcon } />
 	);
 
 	const durationElement =
@@ -90,25 +98,32 @@ export const VideoRow = ( {
 	className = '',
 	checked = false,
 	title,
-	thumbnail: defaultThumbnail,
+	titleAdornment = null,
+	thumbnail,
 	showThumbnail = false,
 	duration,
 	uploadDate,
 	plays,
 	isPrivate,
-	privacySetting,
-	onVideoDetailsClick,
+	onActionClick,
 	onSelect,
-	showEditButton = true,
+	showActionButton = true,
 	showQuickActions = true,
+	showCheckbox = true,
 	loading = false,
+	uploading = false,
+	processing = false,
 	isUpdatingPoster = false,
+	actionButtonLabel = __( 'Edit video details', 'jetpack-videopress-pkg' ),
+	disableActionButton = false,
+	disabled = false,
+	uploadProgress,
 }: VideoRowProps ) => {
 	const textRef = useRef( null );
 	const checkboxRef = useRef( null );
 
 	const [ isSmall ] = useBreakpointMatch( 'sm' );
-	const [ showActions, setShowActions ] = useState( false );
+	const [ showActionsState, setShowActions ] = useState( false );
 	const [ keyPressed, setKeyDown ] = useState( false );
 	const [ expanded, setExpanded ] = useState( false );
 
@@ -117,27 +132,21 @@ export const VideoRow = ( {
 	const isEllipsisActive = textRef?.current?.offsetWidth < textRef?.current?.scrollWidth;
 
 	const showTitleLabel = ! isSmall && isEllipsisActive;
-	const showStats = ( ! showActions && ! isSmall ) || ( isSmall && expanded ) || loading;
-	const showBottom = ! isSmall || ( isSmall && expanded );
-
-	const privacyIsSetToPrivate = privacySetting
-		? VIDEO_PRIVACY_LEVELS[ privacySetting ] === VIDEO_PRIVACY_LEVEL_PRIVATE
-		: false;
-
-	let thumbnail = defaultThumbnail;
-	thumbnail = loading || isUpdatingPoster ? <Placeholder width={ 90 } height={ 50 } /> : thumbnail;
-
+	const showActions =
+		showActionsState && ( showActionButton || showQuickActions ) && ! loading && ! disabled;
+	const showStats = ! loading && ( ( ! isSmall && ! showActions ) || ( isSmall && expanded ) );
+	const showBottom = ! loading && ( ! isSmall || ( isSmall && expanded ) );
 	const canExpand =
 		isSmall &&
 		! loading &&
-		( showEditButton ||
+		( showActionButton ||
 			Boolean( duration ) ||
 			Number.isFinite( plays ) ||
 			typeof isPrivate === 'boolean' );
 
-	const isSpaceOrEnter = code => code === 'Space' || code === 'Enter';
+	const hoverDisabled = isSmall || loading || disabled;
 
-	const editVideoLabel = __( 'Edit video details', 'jetpack-videopress-pkg' );
+	const isSpaceOrEnter = code => code === 'Space' || code === 'Enter';
 
 	const wrapperAriaLabel = sprintf(
 		/* translators: 1 Video title, 2 Video duration, 3 Video upload date */
@@ -155,9 +164,13 @@ export const VideoRow = ( {
 		callback?.( event );
 	};
 
-	const editDetailsButton = (
-		<Button size="small" onClick={ handleClickWithStopPropagation( onVideoDetailsClick ) }>
-			{ editVideoLabel }
+	const actionButton = (
+		<Button
+			size="small"
+			onClick={ handleClickWithStopPropagation( onActionClick ) }
+			disabled={ disableActionButton }
+		>
+			{ actionButtonLabel }
 		</Button>
 	);
 
@@ -202,27 +215,30 @@ export const VideoRow = ( {
 			tabIndex={ 0 }
 			onKeyDown={ isSmall ? null : handleKeyDown }
 			onKeyUp={ isSmall ? null : handleKeyUp }
-			onMouseOver={ isSmall || loading ? null : handleOver }
-			onMouseLeave={ isSmall || loading ? null : handleLeave }
+			onMouseOver={ hoverDisabled ? null : handleOver }
+			onMouseLeave={ hoverDisabled ? null : handleLeave }
 			onClick={ isSmall ? null : handleClick }
 			aria-label={ wrapperAriaLabel }
 			className={ classNames(
 				styles[ 'video-row' ],
 				{
 					[ styles.pressed ]: keyPressed,
+					[ styles.disabled ]: disabled,
 				},
 				className
 			) }
 		>
-			<div className={ classNames( { [ styles[ 'checkbox-wrapper-small' ] ]: isSmall } ) }>
-				<Checkbox
-					ref={ checkboxRef }
-					checked={ checked && ! loading }
-					tabIndex={ -1 }
-					onChange={ onSelect }
-					disabled={ loading }
-				/>
-			</div>
+			{ showCheckbox && (
+				<div className={ classNames( { [ styles[ 'checkbox-wrapper-small' ] ]: isSmall } ) }>
+					<Checkbox
+						ref={ checkboxRef }
+						checked={ checked && ! loading }
+						tabIndex={ -1 }
+						onChange={ onSelect }
+						disabled={ loading }
+					/>
+				</div>
+			) }
 			<div
 				className={ classNames( styles[ 'video-data-wrapper' ], {
 					[ styles.small ]: isSmall,
@@ -236,9 +252,13 @@ export const VideoRow = ( {
 					{ showThumbnail && (
 						<div className={ styles.poster }>
 							<VideoThumbnail
-								isPrivate={ privacyIsSetToPrivate }
 								thumbnail={ thumbnail }
+								loading={ loading || isUpdatingPoster }
+								uploading={ uploading }
+								processing={ processing }
 								blankIconSize={ 28 }
+								uploadProgress={ uploadProgress }
+								isRow
 							/>
 						</div>
 					) }
@@ -252,8 +272,13 @@ export const VideoRow = ( {
 						{ loading ? (
 							<Placeholder height={ 30 } />
 						) : (
-							<Text variant="title-small" className={ styles.title } ref={ textRef }>
+							<Text
+								variant="title-small"
+								className={ classNames( styles.title, { [ styles.disabled ]: disabled } ) }
+								ref={ textRef }
+							>
 								{ title }
+								{ titleAdornment }
 							</Text>
 						) }
 
@@ -271,9 +296,9 @@ export const VideoRow = ( {
 				</div>
 				{ showBottom && (
 					<div className={ classNames( styles[ 'meta-wrapper' ], { [ styles.small ]: isSmall } ) }>
-						{ showActions && ( showEditButton || showQuickActions ) && ! loading && (
+						{ showActions && (
 							<div className={ styles.actions }>
-								{ showEditButton && editDetailsButton }
+								{ showActionButton && actionButton }
 								{ showQuickActions && id && <ConnectVideoQuickActions videoId={ id } /> }
 							</div>
 						) }
@@ -288,7 +313,7 @@ export const VideoRow = ( {
 						) }
 						{ isSmall && (
 							<div className={ styles[ 'mobile-actions' ] }>
-								{ showEditButton && editDetailsButton }
+								{ showActionButton && actionButton }
 								{ showQuickActions && id && <ConnectVideoQuickActions videoId={ id } /> }
 							</div>
 						) }
@@ -300,16 +325,21 @@ export const VideoRow = ( {
 };
 
 export const ConnectVideoRow = ( { id, ...restProps }: VideoRowProps ) => {
-	const { isDeleting, uploading, processing, isUpdatingPoster, data } = useVideo( id );
+	const { isDeleting, uploading, processing, isUpdatingPoster, data, uploadProgress } = useVideo(
+		id
+	);
 	const loading = ( isDeleting || restProps?.loading ) && ! uploading && ! processing;
 	return (
 		<VideoRow
 			id={ id }
 			{ ...restProps }
 			loading={ loading }
+			uploading={ uploading }
 			isUpdatingPoster={ isUpdatingPoster }
+			processing={ processing }
 			showThumbnail
 			privacySetting={ data.privacySetting }
+			uploadProgress={ uploadProgress }
 		/>
 	);
 };