+++ projects/js-packages/licensing/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.6.4 - 2022-11-08
+### Changed
+- Updated package dependencies. [#27289]
+
+## 0.6.3 - 2022-11-01
+### Changed
+- Updated package dependencies. [#27196]
+
 ## 0.6.2 - 2022-10-17
 ### Changed
 - Updated package dependencies. [#26851]