+++ projects/packages/videopress/CHANGELOG.md
@@ -5,6 +5,81 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2022-11-07
+### Added
+- Adding the new Video Chapter block. [#27240]
+- Videopress: add a notice and a modal informing the users how they can create chapters. [#27244]
+- VideoPress: add inspector controls to VideoPress chapters block. [#27242]
+- VideoPress: add Site Settings section. [#27290]
+- VideoPress: add video chapters style selector component. [#27237]
+- VideoPress: add VideoPress video block select control. [#27214]
+- VideoPress: introduce initial layout for chapters. [#27246]
+- VideoPress: scaffolding Video Chapters block. [#27241]
+- VideoPress: show message when there are no search results. [#27275]
+
+### Changed
+- Modified the VideoPress block's prompt to add a VideoChapters block so that the button functions as expected. [#27248]
+- VideoPress: improve the re-rendering process of the video player when editing video props. [#27192]
+- VideoPress: re-organize hooks to keep video data in sync. [#27227]
+
+### Fixed
+- VideoPress: change base number to 10 and total space value to 10^12 so we refer to it in terabytes instead of tebibytes, keeping consistency between marketing and product. [#27274]
+- VideoPress: fix privacy column icon. [#27277]
+- VideoPress: make sure media details are set before using them. [#27262]
+
+## [0.6.5] - 2022-11-01
+### Added
+- VideoPress: add getMediaToken() async helper [#27180]
+- VideoPress: add isBetaExtension() helper [#27179]
+- VideoPress: add isExtensionEnabled() helper" [#27156]
+- VideoPress: Add upload progress to VideoThumbnail and correct states on VideoRow [#27098]
+- VideoPress: expose site-type to the client side [#27191]
+- VideoPress: implement extensions (beta) handling [#27133]
+- VideoPress: migrate video chapters feature from jetpack to videopress [#27178]
+
+### Changed
+- Updated package dependencies. [#27089]
+- VideoPress: Change deprecated prop on VideoQuickActions [#27057]
+- VideoPress: Change the playback token handling to expire tokens after 24h. [#27136]
+- VideoPress: fix width of the actionable placeholder of v6 [#27099]
+
+### Fixed
+- VideoPress: Add processing state to video thumbnail on edit page [#27148]
+- VideoPress: Allow actions and stats on VideoCard when processing video and fix styles [#27123]
+- VideoPress: fix trying to get token when user is disconnected bug [#27067]
+- VideoPress: Fix typescript issues [#27066]
+- VideoPress: Get video title from backend after upload to avoid dirty form while processing [#27188]
+
+## [0.6.4] - 2022-10-25
+### Fixed
+- VideoPress: Update polling time when processing [#27056]
+
+## [0.6.3] - 2022-10-25
+### Added
+- VideoPress: Add component unload prevention on video details edit when there are unsaved changes [#26919]
+- VideoPress: add GlobalNotice component [#26973]
+- VideoPress: add loading state to connect button righ after user clicks on it [#26958]
+- VideoPress: Add support to the use of playback tokens on the details page, so it's possible to see thumbnails on videos that are private. [#26996]
+- VideoPress: Expose VideoPress playback token generation endpoint to enable client display of private videos and thumbnails. [#26974]
+- VideoPress: give access to secondary admins [#26962]
+- VideoPress: handle show admin or pricing based on a local state [#26977]
+- VideoPress: handle UI when setting private video in the quick actions component [#27035]
+- VideoPress: make tracks works [#27040]
+- VideoPress: show an actionable notice when user connection is required [#26986]
+- VideoPress: Upload from library [#26948]
+
+### Changed
+- Updated package dependencies. [#26705]
+- VideoPress: Hide thumbnail actions and open library directly [#27038]
+- VideoPress: show video thumbnails for private videos in the list view [#27024]
+
+### Fixed
+- VideoPress: Add check for site connectivity on video details edit page [#26969]
+- VideoPress: Check for empty poster image on video polling after upload [#27000]
+- VideoPress: enqueue token bridge script when `init` action [#27021]
+- VideoPress: Fix unsaved changes prompt when updating video thumbnail [#26989]
+- VideoPress: Remove ConnectionErrorNotice component on dashboard when there is no connection issue [#26997]
+
 ## [0.6.2] - 2022-10-19
 ### Added
 - VideoPress: Add confirmation before leaving page when upload is in progress [#26912]
@@ -330,6 +405,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created empty package [#24952]
 
+[0.7.0]: https://github.com/Automattic/jetpack-videopress/compare/v0.6.5...v0.7.0
+[0.6.5]: https://github.com/Automattic/jetpack-videopress/compare/v0.6.4...v0.6.5
+[0.6.4]: https://github.com/Automattic/jetpack-videopress/compare/v0.6.3...v0.6.4
+[0.6.3]: https://github.com/Automattic/jetpack-videopress/compare/v0.6.2...v0.6.3
 [0.6.2]: https://github.com/Automattic/jetpack-videopress/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/Automattic/jetpack-videopress/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/Automattic/jetpack-videopress/compare/v0.5.1...v0.6.0